tech(api): move integration test for one entrance into the usecase to the usecase as a whole
